- deployments can timeout and log errors if they take too long
--HG--
extra : convert_revision : svn%3A32cb22c4-c7e1-0310-b164-a889846e9adb/zc.zkdeployment/trunk%4070692

+++ src/zc/zkdeployment/__init__.py
@@ -1,19 +1,34 @@
+import logging
 import os
 import subprocess
 import tempfile
-import logging
+import zc.thread
+
+TIMEOUT_INTERVAL = 900 # seconds
 
 logger = logging.getLogger(__name__)
 
-def run_command(cmd_list):
+
+def run_command(cmd_list, timeout=None):
+    timeout = timeout or TIMEOUT_INTERVAL
     tfile = tempfile.NamedTemporaryFile('w', delete=False)
-    retval = subprocess.call(cmd_list, stdout=tfile, stderr=subprocess.STDOUT)
+    process = []
+    @zc.thread.Thread
+    def worker():
+        process.append(subprocess.Popen(
+            cmd_list, stdout=tfile, stderr=subprocess.STDOUT))
+        process[0].communicate()
+    worker.join(timeout)
+    if worker.is_alive():
+        process[0].terminate()
+        raise RuntimeError('Command failed: ' + ' '.join(cmd_list) +
+                ', took too long')
     tfile.close()
     with open(tfile.name) as f:
         output = f.read()
     os.remove(tfile.name)
-    if retval != 0:
-        logger.error("Comand failed: %r\n  %s" %
+    if process[0].returncode != 0:
+        logger.error("Command failed: %r\n  %s" %
                         (cmd_list,
                          output.replace('\n', '\n  ').strip()))
         raise RuntimeError('Command failed: ' + ' '.join(cmd_list))
